Adding a test script.
Test script will be utilized to check the output of a c program
when ran using a compiler such as gcc. And also checking the output
when running a c program using my c abstract interpreter implementation.
It will then compare the two to see if output is the same.